+++ playlist.go
@@ -74,7 +74,7 @@ func newPlaylist(args Args) *Playlist {
 
 	anko := gomu.anko
 
-	m := anko.getString("music_dir")
+	m := anko.GetString("music_dir")
 	rootDir, err := filepath.Abs(expandTilde(m))
 	if err != nil {
 		err = tracerr.Errorf("unable to find music directory: %e", err)
@@ -89,11 +89,11 @@ func newPlaylist(args Args) *Playlist {
 
 	var rootTextView string
 
-	useEmoji := anko.getBool("use_emoji")
+	useEmoji := anko.GetBool("use_emoji")
 
 	if useEmoji {
 
-		emojiPlaylist := anko.getString("emoji_playlist")
+		emojiPlaylist := anko.GetString("emoji_playlist")
 		rootTextView = fmt.Sprintf("%s %s", emojiPlaylist, path.Base(rootDir))
 
 	} else {
@@ -572,7 +572,7 @@ func ytdl(url string, selPlaylist *tview.TreeNode) error {
 	playlistPath := dir
 	audioPath := extractFilePath(stdout.Bytes(), playlistPath)
 
-	historyPath := gomu.anko.getString("history_path")
+	historyPath := gomu.anko.GetString("history_path")
 
 	err = appendFile(expandTilde(historyPath), url+"\n")
 	if err != nil {
@@ -729,12 +729,12 @@ func (p *Playlist) paste() error {
 }
 
 func setDisplayText(songName string) string {
-	useEmoji := gomu.anko.getBool("use_emoji")
+	useEmoji := gomu.anko.GetBool("use_emoji")
 	if !useEmoji {
 		return songName
 	}
 
-	emojiFile := gomu.anko.getString("emoji_file")
+	emojiFile := gomu.anko.GetString("emoji_file")
 	return fmt.Sprintf(" %s %s", emojiFile, songName)
 }
